+++ includes/Core/ProcessorHelper.php
@@ -6,7 +6,7 @@
  * @package    ApiOpenStudio
  * @subpackage Core
  * @author     john89 (https://gitlab.com/john89)
- * @copyright  2020-2030 ApiOpenStudio
+ * @copyright  2020-2030 Naala Pty Ltd
  * @license    This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
  *             If a copy of the MPL was not distributed with this file,
  *             You can obtain one at https://mozilla.org/MPL/2.0/.
@@ -73,10 +73,10 @@ class ProcessorHelper
     public function isProcessor(&$obj)
     {
         if (is_object($obj)) {
-            return (isset($obj->function) && isset($obj->id));
+            return (isset($obj->processor) && isset($obj->id));
         }
         if (is_array($obj)) {
-            return (isset($obj['function']) && isset($obj['id']));
+            return (isset($obj['processor']) && isset($obj['id']));
         }
         return false;
     }